+++ AWSAppSyncTestCommon/UnitTestHelpers.swift
@@ -7,7 +7,6 @@
 import Foundation
 
 @testable import AWSAppSync
-@testable import AWSAppSyncTestCommon
 
 struct UnitTestHelpers {